🔄️ move DiscordWebhook to classes
