+++ src/classes/DiscordWebhook/interfaces.ts
