+++ src/classes/DiscordWebhook/pricelistUpdate.ts
@@ -5,11 +5,11 @@ import * as timersPromises from 'timers/promises';
 import { UnknownDictionary } from '../../types/common';
 import { Webhook, sendWebhook } from './export';
 
-import log from '../logger';
-import { BuyAndSell } from '../../classes/Pricelist';
-import Options from '../../classes/Options';
+import log from '../../lib/logger';
+import { BuyAndSell } from '../Pricelist';
+import Options from '../Options';
 import { WebhookError } from './utils';
-import { BotInfo } from '../../classes/MyHandler/MyHandler';
+import { BotInfo } from '../MyHandler/MyHandler';
 
 const australiumImageURL: { [defindex: number]: string } = {
     // Australium Ambassador
