+++ src/classes/DiscordWebhook/sendStats.ts
@@ -3,10 +3,10 @@ import pluralize from 'pluralize';
 import SteamID from 'steamid';
 import { sendWebhook } from './utils';
 import { Webhook } from './interfaces';
-import log from '../logger';
+import log from '../../lib/logger';
 import { stats, profit, timeNow } from '../../lib/tools/export';
-import Bot from '../../classes/Bot';
-import loadPollData from '../tools/polldata';
+import Bot from '../Bot';
+import loadPollData from '../../lib/tools/polldata';
 
 export default async function sendStats(bot: Bot, forceSend = false, steamID?: SteamID): Promise<void> {
     const optDW = bot.options.discordWebhook;
