+++ src/classes/MyHandler/offer/processDeclined.ts
@@ -2,7 +2,7 @@ import * as i from '@tf2autobot/tradeoffer-manager';
 import SKU from '@tf2autobot/tf2-sku';
 import Bot from '../../Bot';
 import * as t from '../../../lib/tools/export';
-import sendTradeDeclined from '../../../lib/DiscordWebhook/sendTradeDeclined';
+import sendTradeDeclined from '../../DiscordWebhook/sendTradeDeclined';
 import { KeyPrices } from '../../../classes/Pricelist';
 
 export default function processDeclined(offer: i.TradeOffer, bot: Bot): void {
